WIP: move samples from simulation to own cpp
