+++ python/CMakeLists.txt
@@ -27,6 +27,7 @@ add_library(pyarb MODULE
     profiler.cpp
     pyarb.cpp
     recipe.cpp
+    samples.cpp
     schedule.cpp
     simulation.cpp
     spikes.cpp
