+++ python/simulation.cpp
@@ -10,116 +10,116 @@
 
 namespace pyarb {
 
-// ToDo: move to probes.cpp
-// ToDo: trace entry of different types/container (e.g. vector of doubles to get all samples of a cell)
-
-struct trace_entry {
-    arb::time_type t;
-    double v;
-};
-
-// A helper struct (state) ensuring that only one thread can write to the buffer holding the trace entries (mapped by probe id)
-struct sampler_state {
-    std::mutex mutex;
-    std::unordered_map<arb::cell_member_type, std::vector<trace_entry>> buffer;
-
-    std::vector<trace_entry>& locked_sampler_vec(arb::cell_member_type pid) {
-        // lock the mutex, s.t. other threads cannot write
-        std::lock_guard<std::mutex> lock(mutex);
-        // return or create entry
-        return buffer[pid];
-    }
-
-    // helper function to search probe id in buffer
-    bool has_pid(arb::cell_member_type pid) {
-        std::unordered_map<arb::cell_member_type, std::vector<trace_entry>>::iterator it = buffer.find(pid);
-        return (it != buffer.end());
-    }
-
-    // helper function to push back to locked vector
-    void push_back(arb::cell_member_type pid, trace_entry value) {
-        auto& v = locked_sampler_vec(pid);
-        v.push_back(std::move(value));
-    }
-
-    // helper function to return whole buffer
-    const std::unordered_map<arb::cell_member_type, std::vector<trace_entry>>& samples() const {
-        return buffer;
-    }
-
-    // helper function to return trace entry of probe id
-    const std::vector<trace_entry>& samples_of(arb::cell_member_type pid) {
-        return buffer[pid];
-    }
-};
-
-// A functor that models arb::sampler_function.
-// Holds a shared pointer to the trace_entry used to store the samples, so that if
-// the trace_entry in sample_recorder is garbage collected in Python, stores will
-// not seg fault.
-
-struct sample_callback {
-    std::shared_ptr<sampler_state> sample_store;
-
-    sample_callback(const std::shared_ptr<sampler_state>& state):
-        sample_store(state)
-    {}
-
-    void operator() (arb::cell_member_type probe_id, arb::probe_tag tag, std::size_t n, const arb::sample_record* recs) {
-        // lock before write
-        auto& v = sample_store->locked_sampler_vec(probe_id);
-        for (std::size_t i = 0; i<n; ++i) {
-            if (auto p = arb::util::any_cast<const double*>(recs[i].data)) {
-                v.push_back({recs[i].time, *p});
-            }
-            else {
-                throw std::runtime_error("unexpected sample type");
-            }
-        }
-    };
-};
-
-// Helper type for recording samples from a simulation.
-// This type is wrapped in Python, to expose sample_recorder::sample_store.
-struct sample_recorder {
-    std::shared_ptr<sampler_state> sample_store;
-
-    sample_callback callback() {
-        // initialize the sample_store
-        sample_store = std::make_shared<sampler_state>();
-
-        // The callback holds a copy of sample_store, i.e. the shared
-        // pointer is held by both the sample_recorder and the callback, so if
-        // the sample_recorder is destructed in the calling Python code, attempts
-        // to write to sample_store inside the callback will not seg fault.
-        return sample_callback(sample_store);
-    }
-
-    const std::vector<trace_entry>& samples(arb::cell_member_type pid) const {
-        if (sample_store->has_pid(pid)) {
-            return sample_store->samples_of(pid);
-        }
-        throw std::runtime_error(util::pprintf("probe id {} does not exist", pid));
-    }
-};
-
-// Adds sampler to one probe with pid
-std::shared_ptr<sample_recorder> attach_sample_recorder_on_probe(arb::simulation& sim, arb::time_type interval, arb::cell_member_type pid) {
-    auto r = std::make_shared<sample_recorder>();
-    sim.add_sampler(arb::one_probe(pid), arb::regular_schedule(interval), r->callback());
-    return r;
-}
-
-// Adds sampler to all probes
-std::shared_ptr<sample_recorder> attach_sample_recorder(arb::simulation& sim, arb::time_type interval) {
-    auto r = std::make_shared<sample_recorder>();
-    sim.add_sampler(arb::all_probes, arb::regular_schedule(interval), r->callback());
-    return r;
-}
-
-std::string sample_str(const trace_entry& s) {
-        return util::pprintf("<arbor.sample: time {} ms, value {}>", s.t, s.v);
-}
+//// ToDo: move to probes.cpp
+//// ToDo: trace entry of different types/container (e.g. vector of doubles to get all samples of a cell)
+//
+//struct trace_entry {
+//    arb::time_type t;
+//    double v;
+//};
+//
+//// A helper struct (state) ensuring that only one thread can write to the buffer holding the trace entries (mapped by probe id)
+//struct sampler_state {
+//    std::mutex mutex;
+//    std::unordered_map<arb::cell_member_type, std::vector<trace_entry>> buffer;
+//
+//    std::vector<trace_entry>& locked_sampler_vec(arb::cell_member_type pid) {
+//        // lock the mutex, s.t. other threads cannot write
+//        std::lock_guard<std::mutex> lock(mutex);
+//        // return or create entry
+//        return buffer[pid];
+//    }
+//
+//    // helper function to search probe id in buffer
+//    bool has_pid(arb::cell_member_type pid) {
+//        std::unordered_map<arb::cell_member_type, std::vector<trace_entry>>::iterator it = buffer.find(pid);
+//        return (it != buffer.end());
+//    }
+//
+//    // helper function to push back to locked vector
+//    void push_back(arb::cell_member_type pid, trace_entry value) {
+//        auto& v = locked_sampler_vec(pid);
+//        v.push_back(std::move(value));
+//    }
+//
+//    // helper function to return whole buffer
+//    const std::unordered_map<arb::cell_member_type, std::vector<trace_entry>>& samples() const {
+//        return buffer;
+//    }
+//
+//    // helper function to return trace entry of probe id
+//    const std::vector<trace_entry>& samples_of(arb::cell_member_type pid) {
+//        return buffer[pid];
+//    }
+//};
+//
+//// A functor that models arb::sampler_function.
+//// Holds a shared pointer to the trace_entry used to store the samples, so that if
+//// the trace_entry in sample_recorder is garbage collected in Python, stores will
+//// not seg fault.
+//
+//struct sample_callback {
+//    std::shared_ptr<sampler_state> sample_store;
+//
+//    sample_callback(const std::shared_ptr<sampler_state>& state):
+//        sample_store(state)
+//    {}
+//
+//    void operator() (arb::cell_member_type probe_id, arb::probe_tag tag, std::size_t n, const arb::sample_record* recs) {
+//        // lock before write
+//        auto& v = sample_store->locked_sampler_vec(probe_id);
+//        for (std::size_t i = 0; i<n; ++i) {
+//            if (auto p = arb::util::any_cast<const double*>(recs[i].data)) {
+//                v.push_back({recs[i].time, *p});
+//            }
+//            else {
+//                throw std::runtime_error("unexpected sample type");
+//            }
+//        }
+//    };
+//};
+//
+//// Helper type for recording samples from a simulation.
+//// This type is wrapped in Python, to expose sample_recorder::sample_store.
+//struct sample_recorder {
+//    std::shared_ptr<sampler_state> sample_store;
+//
+//    sample_callback callback() {
+//        // initialize the sample_store
+//        sample_store = std::make_shared<sampler_state>();
+//
+//        // The callback holds a copy of sample_store, i.e. the shared
+//        // pointer is held by both the sample_recorder and the callback, so if
+//        // the sample_recorder is destructed in the calling Python code, attempts
+//        // to write to sample_store inside the callback will not seg fault.
+//        return sample_callback(sample_store);
+//    }
+//
+//    const std::vector<trace_entry>& samples(arb::cell_member_type pid) const {
+//        if (sample_store->has_pid(pid)) {
+//            return sample_store->samples_of(pid);
+//        }
+//        throw std::runtime_error(util::pprintf("probe id {} does not exist", pid));
+//    }
+//};
+//
+//// Adds sampler to one probe with pid
+//std::shared_ptr<sample_recorder> attach_sample_recorder_on_probe(arb::simulation& sim, arb::time_type interval, arb::cell_member_type pid) {
+//    auto r = std::make_shared<sample_recorder>();
+//    sim.add_sampler(arb::one_probe(pid), arb::regular_schedule(interval), r->callback());
+//    return r;
+//}
+//
+//// Adds sampler to all probes
+//std::shared_ptr<sample_recorder> attach_sample_recorder(arb::simulation& sim, arb::time_type interval) {
+//    auto r = std::make_shared<sample_recorder>();
+//    sim.add_sampler(arb::all_probes, arb::regular_schedule(interval), r->callback());
+//    return r;
+//}
+//
+//std::string sample_str(const trace_entry& s) {
+//        return util::pprintf("<arbor.sample: time {} ms, value {}>", s.t, s.v);
+//}
 
 void register_simulation(pybind11::module& m) {
     using namespace pybind11::literals;
@@ -153,32 +153,32 @@ void register_simulation(pybind11::module& m) {
         .def("__str__",  [](const arb::simulation&){ return "<arbor.simulation>"; })
         .def("__repr__", [](const arb::simulation&){ return "<arbor.simulation>"; });
 
-    // Samples
-    pybind11::class_<trace_entry> sample(m, "sample");
-    sample
-        .def(pybind11::init<>())
-        .def_readwrite("time", &trace_entry::t, "The sample time [ms] at a specific probe.")
-        .def_readwrite("value", &trace_entry::v, "The sample value at a specific probe.")
-        .def("__str__",  &sample_str)
-        .def("__repr__", &sample_str);
-
-    // Sample recorder
-    pybind11::class_<sample_recorder, std::shared_ptr<sample_recorder>> samplerec(m, "sample_recorder");
-    samplerec
-        .def(pybind11::init<>())
-        .def("samples", &sample_recorder::samples,
-            "A list of the recorded samples of a probe with probe id.",
-            "pid"_a);
-
-    m.def("attach_sample_recorder", &attach_sample_recorder,
-        "Attach a sample recorder to an arbor simulation.\n"
-        "The recorder will record all samples from a sampling interval [ms] matching all probe ids.",
-        "simulation"_a, "interval"_a);
-
-    m.def("attach_sample_recorder_on_probe", &attach_sample_recorder_on_probe,
-        "Attach a sample recorder to an arbor simulation.\n"
-        "The recorder will record all samples from a sampling interval [ms] matching one probe id.",
-        "simulation"_a, "interval"_a, "pid"_a);
+//    // Samples
+//    pybind11::class_<trace_entry> sample(m, "sample");
+//    sample
+//        .def(pybind11::init<>())
+//        .def_readwrite("time", &trace_entry::t, "The sample time [ms] at a specific probe.")
+//        .def_readwrite("value", &trace_entry::v, "The sample value at a specific probe.")
+//        .def("__str__",  &sample_str)
+//        .def("__repr__", &sample_str);
+//
+//    // Sample recorder
+//    pybind11::class_<sample_recorder, std::shared_ptr<sample_recorder>> samplerec(m, "sample_recorder");
+//    samplerec
+//        .def(pybind11::init<>())
+//        .def("samples", &sample_recorder::samples,
+//            "A list of the recorded samples of a probe with probe id.",
+//            "pid"_a);
+//
+//    m.def("attach_sample_recorder", &attach_sample_recorder,
+//        "Attach a sample recorder to an arbor simulation.\n"
+//        "The recorder will record all samples from a sampling interval [ms] matching all probe ids.",
+//        "simulation"_a, "interval"_a);
+//
+//    m.def("attach_sample_recorder_on_probe", &attach_sample_recorder_on_probe,
+//        "Attach a sample recorder to an arbor simulation.\n"
+//        "The recorder will record all samples from a sampling interval [ms] matching one probe id.",
+//        "simulation"_a, "interval"_a, "pid"_a);
 }
 
 } // namespace pyarb
